fix: refactor codes and fix bugs
