+++ src/js/dom/removebookorchangestatus.js
@@ -15,25 +15,23 @@ const removeBookOrChangeStatus = e => {
     const databaseName = target.dataset.databasename;
     const todos = database.getItems(databaseName);
     currentStates.todelete = 'yes';
-    const id = parseInt(target.id, 10);
-    const newTodos = removeBookLogic(id, todos);
+    const newTodos = removeBookLogic(target.id, todos);
     removeFromUi(target);
     const undo = document.querySelector('.undo');
     undo.classList.remove('d_none');
     undo.addEventListener('click', restore);
-    const undoParams = [target, databaseName, newTodos.todos, undo];
+    const undoParams = [target, databaseName, newTodos, undo];
     undoDelete(...undoParams);
   } else if (target.classList.contains('changeStatus')) {
-    const id = parseInt(target.dataset.identity, 10);
     const databaseName = target.dataset.databasename;
     const todos = database.getItems(databaseName);
-    const newTodos = changeStatusLogic(todos, id);
+    const newTodos = changeStatusLogic(todos, target.dataset.identity);
     changeStatus(...newTodos.changeStatusParams);
     database.setItemToDatabase(databaseName, newTodos.todosList);
   } else if (target.classList.contains('edit')) {
     const editTodoForm = document.querySelectorAll('[data-formtype="edit"]');
     editTodoForm.forEach(fom => {
-      const id = parseInt(fom.dataset.databaseid, 10);
+      const id = fom.dataset.databaseid;
       const fomBtn = fom.querySelector(`#editTodo${id}`);
       fomBtn.addEventListener('click', editTodo);
     });
