+++ src/js/functions/changeStatusLogic.js
@@ -1,6 +1,6 @@
 const changeStatusLogic = (todos, id) => {
   const todosList = todos;
-  let changeStatusParams;
+  let changeStatusParams = [];
   for (let i = 0; i < todosList.length; i += 1) {
     if (todosList[i].id === id) {
       if (todosList[i].status === true) {
