+++ src/js/functions/dynamicid.js
@@ -1,18 +1,12 @@
-import database from '../storage/database';
-
-const dynamicId = databaseName => {
-  let id;
-  if (localStorage.getItem(databaseName) === null) {
-    id = 0;
-  } else {
-    const theLength = localStorage.getItem(databaseName).length;
-    if (theLength < 3) {
-      id = 0;
-    } else {
-      const items = database.getItems(databaseName);
-      id = items[items.length - 1].id + 1;
-    }
-  }
+const dynamicId = () => {
+  const a = '';
+  const c = Date.now() / 1000;
+  let d = c
+    .toString(16)
+    .split('.')
+    .join('');
+  while (d.length < 14) d += '0';
+  const id = a + d;
 
   return id;
 };
