+++ src/js/functions/project.js
@@ -11,7 +11,7 @@ const create = e => {
   const {
     dataset: { formtype: formType, databasename: databaseName },
   } = target;
-  const item = dynamicValues(dynamicId(databaseName), values);
+  const item = dynamicValues(dynamicId(), values);
   addItem(item, databaseName);
   addItemToUi(item)[databaseName][formType]();
   target.reset();
